+++ test/function/utils/format.test.js
@@ -61,7 +61,7 @@ describe('format', function() {
   });
 
   describe('bignumber', function () {
-    var math = mathjs({decimals: 20}); // ensure the precision is 20 digits, the default
+    var math = mathjs({precision: 20}); // ensure the precision is 20 digits, the default
 
     it('should format big numbers', function() {
       assert.equal(math.format(math.bignumber(2).dividedBy(7)), '0.28571428571428571429');